make the ENDOOM support actually build

+++ linuxdoom-1.10/i_system.c
@@ -133,11 +133,11 @@ void I_Quit (void)
     exit(0);
 }
 
-void I_QuitEndoom(unsigned char *endoom, size_t len) {
+void I_QuitEndoom(unsigned char *endoom, int len) {
 
     if (len == 4000) {
-        for (size_t j = 0; j < 25; j++) {
-            for (size_t i = 0; i < 80; i++) {
+        for (int j = 0; j < 25; j++) {
+            for (int i = 0; i < 80; i++) {
                 int c = endoom[2*(80*j + i)];
                 putchar(0 <= c && c <= 127 ? c : ' ');
             }
